requirements.txt
amqp==5.2.0
aniso8601==9.0.1
bcrypt==4.1.2
bidict==0.23.1
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
prompt-toolkit==3.0.43
PyJWT==2.8.0
pymongo==4.6.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
simple-websocket==1.0.0
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0
